backend/requirements.txt
fastapi>=0.99.0,<1.0.0
uvicorn>=0.22.0,<1.0.0
sqlmodel>=0.0.8,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0
aiosqlite>=0.18.0,<1.0.0
psycopg2-binary>=2.9.5,<3.0.0
alembic>=1.9.4,<2.0.0
pydantic[email]>=2.0.0,<3.0.0
pytest>=7.3.1,<8.0.0
pillow>=9.5.0,<10.0.0
faker>=18.9.0,<19.0.0
